Warm up the renderer properly.
